Rename the kythe contrib package to codeanalysis.
We want to tie it to what it's for, not to the relatively
obscure name of a tool which may end up being no more than
an implementation detail.

Also renames the goal from 'kythe' to 'index', for the same
reason.
