+++ contrib/codeanalysis/src/python/pants/__init__.py
@@ -1,0 +1,1 @@
+__import__('pkg_resources').declare_namespace(__name__)
