+++ pants.ini
@@ -27,12 +27,12 @@ pythonpath: [
     "%(buildroot)s/contrib/android/src/python",
     "%(buildroot)s/contrib/avro/src/python",
     "%(buildroot)s/contrib/buildrefactor/src/python",
+    "%(buildroot)s/contrib/codeanalysis/src/python",
     "%(buildroot)s/contrib/cpp/src/python",
     "%(buildroot)s/contrib/errorprone/src/python",
     "%(buildroot)s/contrib/findbugs/src/python",
     "%(buildroot)s/contrib/go/src/python",
     "%(buildroot)s/contrib/jax_ws/src/python",
-    "%(buildroot)s/contrib/kythe/src/python",
     "%(buildroot)s/contrib/mypy/src/python",
     "%(buildroot)s/contrib/node/src/python",
     "%(buildroot)s/contrib/python/src/python",
@@ -49,12 +49,12 @@ backend_packages: +[
     "pants.contrib.android",
     "pants.contrib.avro",
     "pants.contrib.buildrefactor",
+    "pants.contrib.codeanalysis",
     "pants.contrib.cpp",
     "pants.contrib.errorprone",
     "pants.contrib.findbugs",
     "pants.contrib.go",
     "pants.contrib.jax_ws",
-    "pants.contrib.kythe",
     "pants.contrib.scalajs",
     "pants.contrib.mypy",
     "pants.contrib.node",
